The sound system has been rebuilt. Now the ADSR is considered an Envelope. An envelope can be applied (could be nullptr) to a voice. Different adjustments can be done. A new type of soundwave has been created: SquareWave, that is a specif version of PulseWave (with 50% duty cycle).  Now also not all type of waves need to be defined for a voice. All emulations have been adadpted to this big change.
ZX81 Emulation didn't work after the last change in the way the render is managed within this emulation. Now everything is more precise. Even the KONG emulation works better. However there is some problems not weel solved. Probably there is a mistake in the way the memory is read...
The emulation of the chip AY-3-8910 has been improved. This chip can also be used in the emulation of external sound for a Spectrum (something to be done later).